More Refactors
