+++ rmf_fleet_adapter/src/rmf_fleet_adapter/events/GoToPlace.cpp
@@ -18,6 +18,7 @@
 #include "GoToPlace.hpp"
 #include "../project_itinerary.hpp"
 #include "internal_ChopeNegotiator.hpp"
+#include "internal_Utilities.hpp"
 #include "PerformAction.hpp"
 
 #include <cstddef>
@@ -29,37 +30,6 @@
 
 namespace rmf_fleet_adapter {
 namespace events {
-
-//==============================================================================
-std::string wp_name(
-  const agv::RobotContext& context,
-  const rmf_traffic::agv::Plan::Goal& goal)
-{
-  const auto& g = context.planner()->get_configuration().graph();
-  const auto& wp = g.get_waypoint(goal.waypoint());
-  if (wp.name())
-    return *wp.name();
-
-  return "#" + std::to_string(goal.waypoint());
-}
-
-//==============================================================================
-std::string wp_name(const agv::RobotContext& context)
-{
-  const auto& g = context.planner()->get_configuration().graph();
-  const auto& locations = context.location();
-  for (const auto& l : locations)
-  {
-    const auto& wp = g.get_waypoint(l.waypoint());
-    if (wp.name())
-      return *wp.name();
-  }
-
-  if (locations.empty())
-    return "<null>";
-
-  return "#" + std::to_string(locations.front().waypoint());
-}
 
 //==============================================================================
 void GoToPlace::add(rmf_task_sequence::Event::Initializer& initializer)
